clean up and SOC
